New tests
- New certificate file with a bigger key
- Some refactors
